+++ onnx2torch/__init__.py
@@ -1,0 +1,1 @@
+from onnx2torch.converter import convert
