+++ onnx2torch/node_converters/batch_norm.py
@@ -3,12 +3,12 @@ __all__ = []
 import torch
 from torch import nn
 
-from onnx2torch.common import OnnxMapping
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import get_shape_from_value_info
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OnnxMapping
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import get_shape_from_value_info
 
 _BN_CLASS_FROM_SPATIAL_RANK = {
     0: nn.BatchNorm1d,
