+++ onnx2torch/node_converters/cast.py
@@ -4,11 +4,11 @@ import torch
 from onnx import TensorProto
 from torch import nn
 
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import onnx_mapping_from_node
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import onnx_mapping_from_node
 
 # pylint: disable=no-member
 TENSOR_TYPE_TO_TORCH_TYPE = {
