+++ onnx2torch/node_converters/comparisons.py
@@ -3,11 +3,11 @@ __all__ = ['OnnxCompare']
 import torch
 from torch import nn
 
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import onnx_mapping_from_node
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import onnx_mapping_from_node
 
 _TORCH_FUNCTION_FROM_ONNX_TYPE = {
     'Equal': torch.eq,
