+++ onnx2torch/node_converters/constant.py
@@ -5,11 +5,11 @@ from typing import Any
 import torch
 from torch import nn
 
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import onnx_mapping_from_node
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import onnx_mapping_from_node
 
 _CONSTANT_PARSING_MAPPING = {
     'value': lambda x: x.to_torch(),
