+++ onnx2torch/node_converters/gemm.py
@@ -4,11 +4,11 @@ import torch
 import torch.nn.functional as F
 from torch import nn
 
-from onnx2torch.common import OnnxMapping
-from onnx2torch.common import OperationConverterResult
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OnnxMapping
+from onnx2torch.utils.common import OperationConverterResult
 
 
 class OnnxGeneralLinear(nn.Linear):
