+++ onnx2torch/node_converters/logical.py
@@ -6,13 +6,13 @@ import torch
 import torch._C as torch_C
 from torch import nn
 
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import old_style_broadcast
-from onnx2torch.common import onnx_mapping_from_node
-from onnx2torch.custom_export_to_onnx import CustomExportToOnnx
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import old_style_broadcast
+from onnx2torch.utils.common import onnx_mapping_from_node
+from onnx2torch.utils.custom_export_to_onnx import CustomExportToOnnx
 
 _TORCH_FUNCTION_FROM_ONNX_TYPE = {
     'Or': torch.logical_or,
