+++ onnx2torch/node_converters/pow.py
@@ -5,12 +5,12 @@ from typing import Optional
 import torch
 from torch import nn
 
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import old_style_broadcast
-from onnx2torch.common import onnx_mapping_from_node
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import old_style_broadcast
+from onnx2torch.utils.common import onnx_mapping_from_node
 
 
 class OnnxPow(nn.Module):
