+++ onnx2torch/node_converters/reduce.py
@@ -14,14 +14,14 @@ import torch
 import torch._C as torch_C
 from torch import nn
 
-from onnx2torch.common import OnnxMapping
-from onnx2torch.common import OperationConverterResult
-from onnx2torch.common import get_const_value
-from onnx2torch.common import onnx_mapping_from_node
-from onnx2torch.custom_export_to_onnx import CustomExportToOnnx
 from onnx2torch.node_converters.registry import add_converter
 from onnx2torch.onnx_graph import OnnxGraph
 from onnx2torch.onnx_node import OnnxNode
+from onnx2torch.utils.common import OnnxMapping
+from onnx2torch.utils.common import OperationConverterResult
+from onnx2torch.utils.common import get_const_value
+from onnx2torch.utils.common import onnx_mapping_from_node
+from onnx2torch.utils.custom_export_to_onnx import CustomExportToOnnx
 
 
 def _initialize_none_dim(dim: Optional[Union[int, Tuple[int, ...]]], input_dim: int):
