+++ onnx2torch/utils/common.py
